Python String center() Method
Print the word "banana", taking up the space of 20 characters, with "banana" in the middle:

Definition and Usage
The center() method will center align the string, using a specified character (space is default) as the fill character.

Syntax
string.center(length, character)
Parameter Values
Parameter	Description
length	Required. The length of the returned string
character	Optional. The character to fill the missing space on each side. Default is " " (space)