Python String format() Method
Example

Insert the price inside the placeholder, the price should be in fixed point, two-decimal format:

Definition and Usage
The format() method formats the specified value(s) and insert them inside the string's placeholder.

The placeholder is defined using curly brackets: {}. Read more about the placeholders in the Placeholder section below.

The format() method returns the formatted string.

Syntax
string.format(value1, value2...)
Parameter Values
Parameter	Description
value1, value2...	Required. One or more values that should be formatted and inserted in the string.

The values are either a list of values separated by commas, a key=value list, or a combination of both.

The values can be of any data type.